新增文件删除功能，支持通过文件ID或文件对象清除文件并记录日志

+++ app-common/src/main/java/com/eryansky/modules/disk/service/FileService.java
@@ -230,9 +230,48 @@ public class FileService extends CrudService<FileDao, File> {
     public void deleteFileAndClearDisk(String fileId) {
         File file = dao.get(fileId);
         try {
-            dao.delete(file);
             iFileManager.deleteFile(file.getFilePath());
             logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
+            dao.delete(file);
+        } catch (IOException e) {
+            logger.error("删除文件[{}]失败,{}", new Object[]{file.getFilePath(), e.getMessage()});
+        } catch (Exception e) {
+            logger.error(e.getMessage());
+        }
+    }
+
+
+    /**
+     *
+     * 文件删除 不处理引用
+     * @param file 文件
+     */
+    public void deleteFileAndClearDisk(File file) {
+        if(null == file){
+            return;
+        }
+        try {
+            iFileManager.deleteFile(file.getFilePath());
+            logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
+            dao.delete(file);
+        } catch (IOException e) {
+            logger.error("删除文件[{}]失败,{}", new Object[]{file.getFilePath(), e.getMessage()});
+        } catch (Exception e) {
+            logger.error(e.getMessage());
+        }
+    }
+
+    /**
+     *
+     * 文件删除 不处理引用
+     * @param fileId 文件ID
+     */
+    public void clearFileAndClearDisk(String fileId) {
+        File file = dao.get(fileId);
+        try {
+            iFileManager.deleteFile(file.getFilePath());
+            logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
+            dao.clear(file);
         } catch (IOException e) {
             logger.error("删除文件[{}]失败,{}", new Object[]{file.getFilePath(), e.getMessage()});
         } catch (Exception e) {
@@ -245,14 +284,14 @@ public class FileService extends CrudService<FileDao, File> {
      * 文件删除 不处理引用
      * @param file 文件
      */
-    public void deleteFileAndClearDisk(File file) {
+    public void clearFileAndClearDisk(File file) {
         if(null == file){
             return;
         }
         try {
-            dao.delete(file);
             iFileManager.deleteFile(file.getFilePath());
             logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
+            dao.clear(file);
         } catch (IOException e) {
             logger.error("删除文件[{}]失败,{}", new Object[]{file.getFilePath(), e.getMessage()});
         } catch (Exception e) {
